src/app/extensions/cam-quotes/components/quotes-reject-dialog/quotes-reject-dialog.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { RouterTestingModule } from '@angular/router/testing';
import { NgbModalModule } from '@ng-bootstrap/ng-bootstrap';

import { QuotesRejectDialogComponent } from './quotes-reject-dialog.component';

describe('Quotes Reject Dialog Component', () => {
  let component: QuotesRejectDialogComponent;
  let fixture: ComponentFixture<QuotesRejectDialogComponent>;
  let element: HTMLElement;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [QuotesRejectDialogComponent],
      imports: [NgbModalModule, RouterTestingModule],
    }).compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(QuotesRejectDialogComponent);
    component = fixture.componentInstance;
    element = fixture.nativeElement;
  });

  it('should be created', () => {
    expect(component).toBeTruthy();
    expect(element).toBeTruthy();
    expect(() => fixture.detectChanges()).not.toThrow();
  });
});
